clean code and features

+++ src/main/java/com/maxwell/myCollection/service/impl/CategoryServiceImpl.java
@@ -1,14 +1,17 @@
 package com.maxwell.myCollection.service.impl;
 
 import java.util.List;
-import java.util.Optional;
+import java.util.Objects;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.maxwell.myCollection.entity.CategoryEntity;
+import com.maxwell.myCollection.exception.ResourceNotFoundException;
+import com.maxwell.myCollection.model.Category;
 import com.maxwell.myCollection.repository.CategoryRepository;
 import com.maxwell.myCollection.service.CategoryService;
+import com.maxwell.myCollection.utils.CategoryMapper;
 
 @Service
 public class CategoryServiceImpl implements CategoryService {
@@ -17,27 +20,39 @@ public class CategoryServiceImpl implements CategoryService {
 	private CategoryRepository repository;
 
 	@Override
-	public List<CategoryEntity> findAll() {
-		return repository.findAll();
+	public List<Category> findAll() {
+		return CategoryMapper.convertEntitiesToModel(repository.findAll());
 	}
 
 	@Override
-	public Optional<CategoryEntity> findById(Long id) {
-		return repository.findById(id);
+	public Category findById(Long id) {
+		return CategoryMapper.convertEntityToModel(repository.findById(id).orElseThrow());
 	}
 
 	@Override
-	public CategoryEntity addCategory(CategoryEntity category) {
-		return repository.save(category);
+	public Category addCategory(CategoryEntity category) {
+		try {
+			return CategoryMapper.convertEntityToModel(repository.save(category));
+		} catch (Exception e) {
+			throw new ResourceNotFoundException("Something went wrong " + e.getMessage());
+		}
 	}
 
 	@Override
-	public CategoryEntity updateCategory(CategoryEntity category) {
-		return repository.save(category);
+	public Category updateCategory(CategoryEntity category) {
+		try {
+			return CategoryMapper.convertEntityToModel(repository.save(category));
+		} catch (Exception e) {
+			throw new ResourceNotFoundException("Something went wrong " + e.getMessage());
+		}
 	}
 
 	@Override
 	public void removeCategory(Long id) {
+		Category category = CategoryMapper.convertEntityToModel(repository.findById(id).orElseThrow());
+		if (Objects.isNull(category)) {
+			throw new ResourceNotFoundException("Category does not exist");
+		}
 		repository.deleteById(id);
 	}
