+++ src/main/java/com/maxwell/myCollection/service/impl/CommentaryServiceImpl.java
@@ -1,14 +1,17 @@
 package com.maxwell.myCollection.service.impl;
 
 import java.util.List;
-import java.util.Optional;
+import java.util.Objects;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.maxwell.myCollection.entity.CommentaryEntity;
+import com.maxwell.myCollection.exception.ResourceNotFoundException;
+import com.maxwell.myCollection.model.Commentary;
 import com.maxwell.myCollection.repository.CommentaryRepository;
 import com.maxwell.myCollection.service.CommentaryService;
+import com.maxwell.myCollection.utils.CommentaryMapper;
 
 @Service
 public class CommentaryServiceImpl implements CommentaryService {
@@ -17,33 +20,49 @@ public class CommentaryServiceImpl implements CommentaryService {
 	private CommentaryRepository repository;
 
 	@Override
-	public List<CommentaryEntity> findAll() {
-		return repository.findAll();
+	public List<Commentary> findAll() {
+		return CommentaryMapper.convertEntitiesToModel(repository.findAll());
 	}
 
 	@Override
-	public Optional<CommentaryEntity> findById(Long id) {
-		return repository.findById(id);
+	public Commentary findById(Long id) {
+		Commentary commentary = CommentaryMapper.convertEntityToModel(repository.findById(id).orElseThrow());
+		if (commentary == null) {
+			throw new ResourceNotFoundException("Commentary does not exist");
+		}
+		return commentary;
 	}
 
 	@Override
-	public CommentaryEntity addCommentary(CommentaryEntity commentary) {
-		return repository.save(commentary);
+	public Commentary addCommentary(CommentaryEntity commentary) {
+		try {
+			return CommentaryMapper.convertEntityToModel(repository.save(commentary));
+		} catch (Exception e) {
+			throw new ResourceNotFoundException("Something went wrong " + e.getMessage());
+		}
 	}
 
 	@Override
-	public CommentaryEntity updateCommentary(CommentaryEntity commentary) {
-		return repository.save(commentary);
+	public Commentary updateCommentary(CommentaryEntity commentary) {
+		try {
+			return CommentaryMapper.convertEntityToModel(repository.save(commentary));
+		} catch (Exception e) {
+			throw new ResourceNotFoundException("Something went wrong " + e.getMessage());
+		}
 	}
 
 	@Override
 	public void removeCommentary(Long id) {
+		Commentary commentary = CommentaryMapper.convertEntityToModel(repository.findById(id).orElseThrow());
+		if (Objects.isNull(commentary)) {
+			throw new ResourceNotFoundException("Commentary does not exist");
+		}
 		repository.deleteById(id);
 	}
 
 	@Override
-	public List<CommentaryEntity> findByItemId(Long id) {
-		return repository.findByItemId(id);
+	public List<Commentary> findByItemId(Long id) {
+		return CommentaryMapper.convertEntitiesToModel(repository.findByItemId(id));
 	}
 
 }
